forge fmt

+++ foundry/src/exercises/Router.sol
@@ -90,8 +90,8 @@ contract Router is TStore, IUnlockCallback {
     {
         uint256 action = _getAction();
         // Write your code here
-        if (action == SWAP_EXACT_IN_SINGLE){
-            (address msgSender, ExactInputSingleParams memory params) = 
+        if (action == SWAP_EXACT_IN_SINGLE) {
+            (address msgSender, ExactInputSingleParams memory params) =
                 abi.decode(data, (address, ExactInputSingleParams));
             (int128 amount0, int128 amount1) = _swap(
                 params.poolKey,
@@ -126,12 +126,13 @@ contract Router is TStore, IUnlockCallback {
                 amountOut: amountOut
             });
             return abi.encode(amountOut);
-        }else if (action == SWAP_EXACT_OUT_SINGLE){
-            (address msgSender, ExactOutputSingleParams memory params) = abi.decode(data, (address, ExactOutputSingleParams));
+        } else if (action == SWAP_EXACT_OUT_SINGLE) {
+            (address msgSender, ExactOutputSingleParams memory params) =
+                abi.decode(data, (address, ExactOutputSingleParams));
             (int128 amount0, int128 amount1) = _swap(
                 params.poolKey,
                 params.zeroForOne,
-                params.amountOut.toInt256(),   // 在swap的内部，通过判断amountSpecified的正负，来断定指定金额是输入还是输出
+                params.amountOut.toInt256(), // 在swap的内部，通过判断amountSpecified的正负，来断定指定金额是输入还是输出
                 params.hookData
             );
             (
@@ -162,16 +163,17 @@ contract Router is TStore, IUnlockCallback {
                 amountOut: amountOut
             });
             return abi.encode(amountIn);
-        }else if(action == SWAP_EXACT_IN){
-            (address msgSender, ExactInputParams memory params) =  abi.decode(data, (address, ExactInputParams));
-            
+        } else if (action == SWAP_EXACT_IN) {
+            (address msgSender, ExactInputParams memory params) =
+                abi.decode(data, (address, ExactInputParams));
+
             uint256 n = params.path.length;
             address currencyIn = params.currencyIn;
             int256 amountIn = params.amountIn.toInt256();
-            for (uint256 i=0; i< n; i++){
+            for (uint256 i = 0; i < n; i++) {
                 PathKey memory path = params.path[i];
-                (address currency0, address currency1) = path.currency 
-                    < currencyIn 
+                (address currency0, address currency1) = path.currency
+                    < currencyIn
                     ? (path.currency, currencyIn)
                     : (currencyIn, path.currency);
                 PoolKey memory key = PoolKey({
@@ -192,45 +194,48 @@ contract Router is TStore, IUnlockCallback {
                 amountIn = (zeroForOne ? amount1 : amount0).toInt256();
             }
             require(
-                uint256(amountIn) >= uint256(params.amountOutMin), "amount out < min"
+                uint256(amountIn) >= uint256(params.amountOutMin),
+                "amount out < min"
             );
             _takeAndSettle({
                 dst: msgSender,
                 currencyIn: params.currencyIn,
-                currencyOut: currencyIn,      // 循环的最后一个
+                currencyOut: currencyIn, // 循环的最后一个
                 amountIn: params.amountIn,
                 amountOut: uint256(amountIn)
             });
             return abi.encode(uint256(amountIn));
-
-            
-        }else if(action == SWAP_EXACT_OUT){
-            (address msgSender, ExactOutputParams memory params) = abi.decode(data, (address, ExactOutputParams));
+        } else if (action == SWAP_EXACT_OUT) {
+            (address msgSender, ExactOutputParams memory params) =
+                abi.decode(data, (address, ExactOutputParams));
             uint256 n = params.path.length;
             address currencyOut = params.currencyOut;
             int256 amountOut = params.amountOut.toInt256();
-            for (uint256 i = n; i> 0; i--){
-                PathKey memory path = params.path[i-1];
-                (address currency0, address currency1) = path.currency < currencyOut
+            for (uint256 i = n; i > 0; i--) {
+                PathKey memory path = params.path[i - 1];
+                (address currency0, address currency1) = path.currency
+                    < currencyOut
                     ? (path.currency, currencyOut)
                     : (currencyOut, path.currency);
                 PoolKey memory key = PoolKey({
                     currency0: currency0,
                     currency1: currency1,
                     fee: path.fee,
-                    tickSpacing:path.tickSpacing,
+                    tickSpacing: path.tickSpacing,
                     hooks: path.hooks
                 });
 
                 bool zeroForOne = currencyOut == currency1;
-                (int128 amount0, int128 amount1) = _swap(
-                    key, zeroForOne, amountOut, path.hookData
-                );
+                (int128 amount0, int128 amount1) =
+                    _swap(key, zeroForOne, amountOut, path.hookData);
                 currencyOut = path.currency;
                 amountOut = (zeroForOne ? -amount0 : -amount1).toInt256();
             }
 
-            require(uint256(amountOut) <= uint(params.amountInMax), "amount in > max");
+            require(
+                uint256(amountOut) <= uint256(params.amountInMax),
+                "amount in > max"
+            );
             _takeAndSettle({
                 dst: msgSender,
                 currencyIn: currencyOut,
@@ -240,7 +245,6 @@ contract Router is TStore, IUnlockCallback {
             });
             return abi.encode(uint256(amountOut));
         }
-
 
         revert UnsupportedAction(action);
     }
@@ -274,7 +278,7 @@ contract Router is TStore, IUnlockCallback {
         currencyIn.transferIn(msg.sender, params.amountInMax);
         poolManager.unlock(abi.encode(msg.sender, params));
         uint256 refunded = _refund(currencyIn, msg.sender);
-        if (refunded < params.amountInMax){
+        if (refunded < params.amountInMax) {
             return params.amountInMax - refunded;
         }
         return 0;
@@ -308,11 +312,10 @@ contract Router is TStore, IUnlockCallback {
         currencyIn.transferIn(msg.sender, params.amountInMax);
         poolManager.unlock(abi.encode(msg.sender, params));
         uint256 refunded = _refund(currencyIn, msg.sender);
-        if (refunded < params.amountInMax){
+        if (refunded < params.amountInMax) {
             return params.amountInMax - refunded;
         }
         return 0;
-
     }
 
     function _swap(
@@ -320,9 +323,9 @@ contract Router is TStore, IUnlockCallback {
         bool zeroForOne,
         int256 amountSpecified,
         bytes memory hookData
-    ) private returns (int128 amount0, int128 amount1){
+    ) private returns (int128 amount0, int128 amount1) {
         int256 d = poolManager.swap({
-            key:key,
+            key: key,
             params: SwapParams({
                 zeroForOne: zeroForOne,
                 amountSpecified: amountSpecified,
@@ -340,20 +343,20 @@ contract Router is TStore, IUnlockCallback {
         address currencyOut,
         uint256 amountIn,
         uint256 amountOut
-    )private {
+    ) private {
         poolManager.take(currencyOut, dst, amountOut);
         poolManager.sync(currencyIn);
-        if (currencyIn == address(0)){
+        if (currencyIn == address(0)) {
             poolManager.settle{value: amountIn}();
-        }else{
+        } else {
             IERC20(currencyIn).transfer(address(poolManager), amountIn);
             poolManager.settle();
         }
     }
 
-    function _refund(address currency, address dst) private returns(uint256){
+    function _refund(address currency, address dst) private returns (uint256) {
         uint256 bal = currency.balanceOf(address(this));
-        if(bal > 0){
+        if (bal > 0) {
             currency.transferOut(dst, bal);
         }
         return bal;
